# [10.x] Refactored LazyCollection::except() to save memory