@taylorotwell 2023-09-26
I think this may be a breaking change? Normal collections use `Arr::forget` under the hood for except, which supports "dot" notation.