@JosephSilber 2023-09-26
> I think this may be a breaking change? Normal collections use `Arr::forget` under the hood for except, which supports "dot" notation.

Oh right!

When this PR came in, I was racking my brain trying to remember why I didn't do it originally.

Now that you've mentioned it, I remember that this was indeed the reason. Didn't think at the time that it's worth reimplementing it all here.

---

If we do want to do it now, we have to decide whether we want to support dot notation for when the children (or other descendants) are also lazy collections 🤔