@fuwasegu 2023-09-27
@taylorotwell @JosephSilber 
thank you for your comment!

Certainly, the Normal Collection supports nested access by dot.
Taylor is right, my changes are disruptive.

The choices are
- Make this change from 11.x and allow destructive changes.
- Give up being lazy and passthru as before
- Support Nested access by dot

And Joseph is right, LazyCollection<LazyCollection> does not support nested access even in its current state, so we will have to think about whether to support it.