@fuwasegu 2023-09-27
I've thought about it a lot, but I couldn't come up with an implementation of except() that supports dot nesting in LazyCollection...

For example, if the keys are ['foo.bar', 'bar', 'foo.baz'], etc., LazyCollection cannot run backwards, so it will have to go around once for each key verification