@fuwasegu 2023-09-27
@morloderex  nice idea !
thanks!