@fuwasegu 2023-09-27
How about this one ?

```php
public function except($keys)
{
    $keys = match (true) {
        $keys instanceof Enumerable => $keys->all(),
        is_array($keys) => $keys,
        is_null($keys) => [],
        default => func_get_args(),
    };

    if (count($keys) === 0) {
        return $this;
    }

    $keys = Arr::undot(array_fill_keys($keys, true));

    $forget = function (LazyCollection|ArrayAccess|array &$data, array $keys) use (&$forget) {
        if ($data instanceof LazyCollection) {
            return new static(function () use ($data, $keys, $forget) {
                foreach ($data as $key => $value) {
                    $shouldYield = true;
                    foreach ($keys as $k => $v) {
                        if (
                            $k === $key
                            && is_array($v)
                            && (
                                $value instanceof LazyCollection
                                || is_array($value)
                                || $value instanceof ArrayAccess
                            )
                        ) {
                            yield $key => $forget($value, $v);
                            $shouldYield = false;
                            break;
                        } elseif ($k === $key) {
                            $shouldYield = false;
                            break;
                        }
                    }
                    if ($shouldYield) {
                        yield $key => $value;
                    }
                }
            });
        } else {
            foreach ($keys as $k => $v) {
                if (is_array($v)) {
                    return $forget($data, $v);
                } else {
                    if (isset($data[$k])) {
                        unset($data[$k]);
                        return $data;
                    }
                }
            }
        }
    };

    return new static(function () use ($keys, $forget) {
        yield from $forget($this, $keys);
    });
}
```